backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
# Blockchain/Ethereum stack (compatible versions)
web3==5.31.4
eth-typing==2.2.2
eth-utils==1.10.0
eth-keys==0.4.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.25.2
pytest-cov==4.1.0
email-validator==2.1.0